fix: Set IAST Instrumented metrics at runtime, nor at request time
